requirements.txt
absl-py==0.13.0
appdirs==1.4.4
asgiref==3.4.1
audioread==2.1.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cycler==0.10.0
decorator==5.0.9
fastapi==0.67.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
grpcio==1.39.0
h11==0.12.0
idna==3.2
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.2
nnpack==0.1.0
numba==0.53.1
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
pooch==1.4.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
SoundFile==0.10.3.post1
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tf==1.0.0
threadpoolctl==2.2.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
Werkzeug==2.0.1
